Course 04 - HW 004 - Logging and configuration in the app.
