+++ src/main/java/com/skypro/course_03/controllers/StudentController.java
@@ -2,15 +2,12 @@ package com.skypro.course_03.controllers;
 
 import com.skypro.course_03.entity.Faculty;
 import com.skypro.course_03.entity.Student;
-import com.skypro.course_03.services.AvatarService;
 import com.skypro.course_03.services.StudentService;
+import com.skypro.course_03.services.StudentServiceImpl;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
 
-import java.io.IOException;
 import java.util.Collection;
 import java.util.Optional;
 
@@ -21,7 +18,7 @@ public class StudentController {
 
     private StudentService studentService;
 
-    public StudentController(StudentService studentService) {
+    public StudentController(StudentServiceImpl studentService) {
         this.studentService = studentService;
     }
