+++ src/main/resources/application.properties
@@ -6,3 +6,8 @@ spring.jpa.hibernate.ddl-auto=update
 students.avatar.dir.path=./uploads
 
 spring.liquibase.change-log=classpath:liquibase/changelog-master.yml
+
+logging.level.com.skypro.course_03.services=DEBUG
+logging.file.name=./hogwarts.txt
+
+server.port=8080
